+++ Trabalho de Implementação/Módulo 1/merge_sort_concorrente.c
@@ -13,14 +13,14 @@ typedef struct {
     int fim;
 } threadArgs;
 
-int *merge_sort_concorrente(int tamanho_vetor, int *vetor, int qtd_threads);
+int *merge_sort_concorrente(long long int tamanho_vetor, int *vetor, int qtd_threads);
 void merge_sort(int *vetor, int inicio, int fim);
 void merge(int *vetor, int inicio, int meio, int fim);
 void *tarefa(void *arg);
-void imprime_vetor(int tamanho_vetor, int *vetor);
+void imprime_vetor(long long int tamanho_vetor, int *vetor);
 
 
-int *merge_sort_concorrente(int tamanho_vetor, int *vetor, int qtd_threads){
+int *merge_sort_concorrente(long long int tamanho_vetor, int *vetor, int qtd_threads){
     pthread_t *thread_id;
     threadArgs *arg;
     threadArgs *vetorThreadArgs[qtd_threads];
@@ -170,7 +170,7 @@ void *tarefa(void *arg){
     pthread_exit(NULL);
 }
 
-void imprime_vetor(int tamanho_vetor, int *v){
+void imprime_vetor(long long int tamanho_vetor, int *v){
     for (int i=0; i<tamanho_vetor; i++){
         printf("%d ", v[i]);
     }
